refactor: conversationEventsGateway의 긴 logic을 service로 분리
