+++ src/conversation-events/conversation-events.gateway.ts
@@ -9,8 +9,6 @@ import {
 } from '@nestjs/websockets';
 import { Server } from 'socket.io';
 import {
-  copyFromBeforeSocket,
-  disconnectBeforeSocket,
   initRoomSocket,
   RoomSocket,
   SocketStatus,
@@ -21,12 +19,10 @@ import {
 } from './conversation-events.filter';
 import { AuthService } from 'src/auth/auth.service';
 import { UsersService } from 'src/users/users.service';
-import { Handshake } from 'socket.io/dist/socket-types';
-import { JwtPayloadDto } from 'src/auth/dto/jwt-payload';
 import { RoomsService } from 'src/rooms/rooms.service';
 import { User } from 'src/users/entities/user.entity';
 import { Room, RoomStatus } from 'src/rooms/room';
-import { SocketInformation } from './interfaces/socket-information.interface';
+import { ConversationEventsService } from './conversation-events.service';
 
 @UseFilters(ConversationEventsFilter)
 @WebSocketGateway(3001, {
@@ -44,6 +40,7 @@ export class ConversationEventsGateway
     private authService: AuthService,
     private usersService: UsersService,
     private roomsService: RoomsService,
+    private conversationEventsService: ConversationEventsService,
   ) {}
 
   @WebSocketServer() server: Server;
@@ -58,8 +55,10 @@ export class ConversationEventsGateway
     let user: User;
     let room: Room;
     try {
-      ({ user, room } = await this.getUserAndRoom(client.handshake));
-      await this.setClientStatus(client, user, room);
+      ({ user, room } = await this.conversationEventsService.getUserAndRoom(
+        client.handshake,
+      ));
+      await this.conversationEventsService.setClientStatus(client, user, room);
     } catch (error) {
       // socket 연결에 실패하는 경우
       this.logger.log(
@@ -71,7 +70,11 @@ export class ConversationEventsGateway
       return;
     }
     initRoomSocket(client, user, room);
-    await this.joinClientInRoom(room, client);
+    await this.conversationEventsService.joinClientInRoom(
+      room,
+      this.server,
+      client,
+    );
 
     this.logger.log(`연결된 Client id: ${client.id} status: ${client.status}`);
   }
@@ -138,106 +141,5 @@ export class ConversationEventsGateway
       }
       room.watingSockets.splice(indexToRemove);
     }
-  }
-
-  private async setClientStatus(client: RoomSocket, user: User, room: Room) {
-    // 이전 연결을 유지한 채로 다시 참여하는 경우
-    const beforeSocket = await this.roomsService.findRoomSocket(room, user);
-    if (beforeSocket) {
-      copyFromBeforeSocket(beforeSocket, client);
-      disconnectBeforeSocket(beforeSocket);
-      return;
-    }
-    switch (room.status) {
-      // 방장이 입장하지 않은 방인 경우
-      case RoomStatus.WATING:
-        if (room.creatorPk != user.pk) {
-          throw new Error('방장이 입장하지 않습니다.');
-        }
-        client.status = SocketStatus.CREATOR;
-        break;
-      // 방장이 초대 중인 경우
-      case RoomStatus.INVITING:
-        client.status = SocketStatus.WAITER;
-        break;
-      case RoomStatus.CREATOR_OUT:
-      case RoomStatus.PARTICIPANT_OUT:
-        let beforInformation: SocketInformation;
-        if (room.status == RoomStatus.CREATOR_OUT) {
-          beforInformation = room.creatorInformation;
-        } else {
-          beforInformation = room.participantInformation;
-        }
-
-        if (beforInformation.userPk != user.pk) {
-          throw new Error(
-            `이미 개최중이거나 종료중인 방입니다: ${room.status}`,
-          );
-        }
-        beforInformation.clearTimeout();
-        beforInformation.setSocket(client);
-        room.status = RoomStatus.RUNNING;
-        break;
-      default:
-        throw new Error(`이미 개최중이거나 종료중인 방입니다: ${room.status}`);
-    }
-  }
-
-  private async joinClientInRoom(room: Room, client: RoomSocket) {
-    if (client.status == SocketStatus.CREATOR) {
-      // 처음 방장이 입장한 경우
-      if (room.status == RoomStatus.WATING) {
-        room.status = RoomStatus.INVITING;
-        this.logger.log(`${room.uuid}에서 초대를 시작합니다`);
-      }
-      room.creatorSocket = client;
-      client.join(room.uuid);
-      this.logger.log('방장이 되었습니다.');
-    } else if (client.status == SocketStatus.PARTICIPANT) {
-      room.participantSocket = client;
-      this.logger.log(`참가자가 되었습니다.`);
-    } else if (client.status == SocketStatus.WAITER) {
-      room.watingSockets.push(client);
-      this.server.to(room.uuid).emit('join-request-by', {
-        message: '참가 요청이 왔습니다.',
-        data: {
-          participant: {
-            name: client.user.name,
-            email: client.user.email,
-          },
-        },
-      });
-      this.logger.log('대기자에 추가되었습니다.');
-    } else {
-      this.logger.error(`동일한 connect 요청 과정에서 예상하지 못 한 에러`);
-      this.logger.error(`${client.status}`);
-      throw new WsException(
-        '동일한 connect 요청 과정에서 문제가 생겼습니다. 백앤드를 불러주세요.',
-      );
-    }
-  }
-
-  private async getUserAndRoom(handshake: Handshake) {
-    const cookie = handshake.headers.cookie;
-    if (cookie == undefined) {
-      throw new Error('로그인 하지 않았습니다');
-    }
-    const cookieMap = new URLSearchParams(cookie.replace(/; /g, '&'));
-    const token: string = cookieMap.get('token');
-    const payload: JwtPayloadDto = await this.authService.getUser(token);
-    const user = await this.usersService.findUserbyPayload(payload);
-
-    let roomUuid = handshake.query.roomUuid;
-    if (Array.isArray(roomUuid)) {
-      roomUuid = roomUuid[0];
-    }
-    this.logger.log(`유저 정보 ${user.email} ${user.name}`);
-    this.logger.log(`받은 room uuid ${roomUuid}`);
-    const room = await this.roomsService.findRoombyUuid(roomUuid);
-    if (!room) {
-      throw new Error('방이 존재하지 않습니다.');
-    }
-
-    return { user, room };
   }
 }
